test: headerdump cli tests
